feat: Add image conversion module with batch processing - Implement image converter supporting batch processing for multiple image formats - Add support for JPEG, PNG, and WebP format conversions - Include quality adjustment options for compressed formats
