--- config.json
@@ -1,1 +1,0 @@
-{"language": "简体中文"}
