+++ locales/zh_CN.json
@@ -1,4 +1,14 @@
 {
+    "title": "花花视频格式批量转换工具",
+    "github_text": "项目GitHub主页",
+    "language_label": "应用程序语言:",
+    "save_settings": "保存设置",
+    "settings_saved": "设置已保存",
+    "settings_saved_msg": "您的设置已成功保存。",
+    "tab_video": "视频",
+    "tab_audio": "音频 (开发中)",
+    "tab_image": "图像",
+    "tab_settings": "设置",
     "input_folder_label": "输入文件夹:",
     "output_folder_label": "输出文件夹:",
     "target_format_label": "目标格式:",
@@ -26,22 +36,29 @@
     "no_video_files_msg": "未找到视频文件在目录: ",
     "conversion_start": "正在查找视频文件...",
     "conversion_start_log": "开始扫描视频文件...",
-    "video_files_found": "找到 {} 个视频文件",
-    "using_threads": "使用 {} 个线程进行转换",
-    "target_format": "目标格式: {}",
-    "output_directory": "输出目录: {}",
+    "video_files_found": "找到 {video_files_count} 个视频文件",
+    "using_threads": "使用 {thread_count} 个线程进行转换",
+    "target_format": "目标格式: {target_format}",
+    "output_directory": "输出目录: {output_directory}",
     "converting": "正在转换...",
     "conversion_complete": "转换完成",
-    "conversion_complete_msg": "转换完成！共处理 {} 个文件，成功 {} 个，失败 {} 个。",
+    "conversion_complete_msg": "转换完成！共处理 {total_files} 个文件，成功 {success_files} 个，失败 {failed_files} 个。",
     "conversion_failed": "转换失败",
     "conversion_failed_msg": "转换过程中发生错误: ",
-    "file_conversion_start": "开始转换: {}",
-    "file_conversion_success": "成功: {}",
-    "file_conversion_error": "错误: 处理 {} 时出错 - {}",
-    "file_conversion_skip": "跳过: {} 已存在",
-    "file_not_found": "错误: {} 不是有效的文件",
-    "unknown_error": "未知错误: 转换 {} 时发生异常 - {}",
+    "file_conversion_start": "开始转换: {file_path}",
+    "file_conversion_success": "成功: {file_path}",
+    "file_conversion_error": "错误: 处理 {file_path} 时出错 - {error_msg}",
+    "file_conversion_skip": "跳过: {file_path} 已存在",
+    "file_not_found": "错误: {file_path} 不是有效的文件",
+    "unknown_error": "未知错误: 转换 {file_path} 时发生异常 - {error_msg}",
     "conversion_finished_title": "完成",
-    "conversion_finished_msg": "转换完成！\n成功: {}\n失败: {}",
-    "set_output_folder": "自动设置输出文件夹为: {}"
+    "conversion_finished_msg": "转换完成！\n成功: {success_files}\n失败: {failed_files}",
+    "set_output_folder": "自动设置输出文件夹为: {output_folder}",
+    "convert_button": "转换",
+    "input_format_label": "输入格式:",
+    "output_format_label": "输出格式:",
+    "error_no_io_folders": "请选择输入和输出文件夹。",
+    "conversion_success": "转换成功完成。",
+    "conversion_failed": "转换失败",
+    "jpeg_quality_label": "JPEG 质量:"
 }
